requirements.txt
alembic==1.15.2
annotated-types==0.7.0
MarkupSafe==3.0.2
psycopg2-binary==2.9.10
pydantic==2.11.3
pydantic_core==2.33.1
pytest==8.3.5
python-dotenv==1.1.0
python-telegram-bot[job-queue]==22.0
requests==2.32.3
SQLAlchemy==2.0.40
Jinja2==3.1.6
click==8.2.1
cryptography~=45.0.4